requirements.txt
# Sage Intacct & API
requests>=2.31.0
xmltodict>=0.13.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Fuzzy Matching
rapidfuzz>=3.5.0
jellyfish>=1.0.0

# Market Data Providers
yfinance>=0.2.33
fredapi>=0.5.1
intrinio-sdk>=6.27.0  # Optional: premium financial data

# Database
sqlalchemy>=2.0.0
sqlite3  # built-in

# Reporting
openpyxl>=3.1.0
jinja2>=3.1.0

# CLI
click>=8.1.0
rich>=13.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Web API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Authentication
PyJWT>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0